Refactor API schemas: move request/response models to separate files for barcode, qrcode, shorturl, and users; update imports in API files. Enhance main.py to handle database startup errors gracefully.
