+++ app/api/qrcode.py
@@ -1,7 +1,7 @@
 import base64
 from fastapi import APIRouter, HTTPException, status, Depends, Body, Response
 from fastapi.responses import RedirectResponse
-from pydantic import BaseModel, Field, AnyUrl
+from pydantic import Field, AnyUrl
 from sqlalchemy.orm import Session
 from sqlalchemy.exc import IntegrityError
 from app.models.models import Qrcode
@@ -31,20 +31,7 @@ db_dependency = Annotated[Session, Depends(get_db)]
 user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
-class QRCodeRequest(BaseModel):
-	original_url: AnyUrl
-	title: str = Field(None, max_length=256)
-	description: Optional[str] = None
-
-class QRCodeResponse(BaseModel):
-	original_url: str
-	qr_code_id: str
-	qr_code_image: str
-	title: str = None
-	description: Optional[str] = None
-	scans: int
-	user_id: int = None
-	created_at: str
+from app.schemas.qrcode import QRCodeRequest, QRCodeResponse
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
