+++ app/main.py
@@ -6,10 +6,18 @@ from fastapi.staticfiles import StaticFiles
 from fastapi.responses import RedirectResponse
 from starlette.middleware.sessions import SessionMiddleware
 from app.core.config import settings
+from sqlalchemy.exc import OperationalError
 
 app = FastAPI()
 
-Base.metadata.create_all(bind=engine) 
+# Create tables on startup, but don't crash the server if DB is down
+@app.on_event("startup")
+async def startup_db():
+	try:
+		Base.metadata.create_all(bind=engine)
+	except OperationalError as e:
+		# Log a concise warning; endpoints that hit DB will still raise 500 until DB is up
+		print("[Startup] Database not available; will retry on first DB access.")
 
 app.add_middleware(SessionMiddleware, secret_key=settings.secret_key)
