+++ app/schemas/__init__.py
@@ -1,0 +1,1 @@
+# Pydantic schemas package
