+++ app/schemas/users.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class UserVerification(BaseModel):
+    password: str
+    new_password: str
